add configure manage system
增加了配置管理系统
