+++ Application/Admin/View/Config/modConfig.html
@@ -29,12 +29,16 @@
 						<div class="am-u-sm-2 am-text-right">备注</div>
 						<div class="am-u-sm-6 am-u-end"><input type="text" class="am-input-sm" name="comment" value="{$data.comment}"></div>
 					</div>
-
 					<div class="am-g am-margin-top-sm">
 						<div class="am-u-sm-2 am-text-right">value</div>
 						<div class="am-u-sm-6 am-u-end">
 							<textarea name="value" rows="" cols="">{$data.value}</textarea>
 						</div>
+					</div>
+					<div class="am-g am-margin-top-sm">
+						<div class="am-u-sm-2 am-text-right">属性</div>
+						<div class="am-u-sm-6"><input type="text" class="am-input-sm" name="attr" value="{$data.attr}"></div>
+						<div class="am-u-sm-4"><p>int<br />0:必须<br />1(default):可擦</p></div>
 					</div>
 
 					<div class="am-g am-margin-top">
