+++ Application/Admin/View/Index/index.html
@@ -5,7 +5,7 @@
 <div class="admin-content">
 
 <div class="am-cf am-padding">
-	<div class="am-fl am-cf"><strong class="am-text-primary am-text-lg">基本信息</strong> / <small>Base Information</small>
+	<div class="am-fl am-cf"><strong class="am-text-primary am-text-lg">站点信息</strong> / <small>Site Information</small>
 	</div>
 </div>
 <ul class="am-avg-sm-1 am-avg-md-4 am-margin am-padding am-text-center admin-content-list ">
